% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cgss2005}
\alias{cgss2005}
\title{Chinese General Social Survey 2005.}
\format{A data frame with 10372 rows and 504 variables:
\describe{
  Variables Introduction in Chinese version: data(codebook2005_cn)
  Variables Introduction in English version: data(codebook2005_en)
}}
\source{
\url{http://cnsda.ruc.edu.cn}{Chinese National Survey Data Archive}

\url{http://cgss.ruc.edu.cn}{Chinese General Social Survey Project}
}
\usage{
cgss2005
}
\description{
Chinese General Social Survey 2005.
}
\keyword{datasets}
