% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codebook2005_en}
\alias{codebook2005_en}
\title{Codebook for CGSS 2005 in English.}
\format{A data frame with 504 rows and 3 variables:
\describe{
  \item{index}{index of this variable in cgss2005 dataset}
  \item{question}{survey question of this variable}
  \item{label_en}{meaning of the answer label in the survey}
}}
\source{
\url{http://cnsda.ruc.edu.cn}{Chinese National Survey Data Archive}

\url{http://cgss.ruc.edu.cn}{Chinese General Social Survey Project}

\url{https://cloud.google.com/translate/docs/}{Google Translation API}
}
\usage{
codebook2005_en
}
\description{
Codebook for CGSS 2005 in English.
}
\keyword{datasets}
