% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codebook2008_cn}
\alias{codebook2008_cn}
\title{Codebook for CGSS 2008 in Chinese.}
\format{A data frame with 1504 rows and 3 variables:
\describe{
  \item{index}{index of this variable in cgss2008 dataset}
  \item{question}{survey question of this variable}
  \item{label_ch}{meaning of the answer label in the survey}
}}
\source{
\url{http://cnsda.ruc.edu.cn}{Chinese National Survey Data Archive}

\url{http://cgss.ruc.edu.cn}{Chinese General Social Survey Project}
}
\usage{
codebook2008_cn
}
\description{
Codebook for CGSS 2008 in Chinese.
}
\keyword{datasets}
