% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_codebook.R
\name{codebook_trans}
\alias{codebook_trans}
\title{Use Google Translation API to translate all the
question and answer label into English.
Build Chinese General Social Survey codebook in both
English and Chinese version.}
\source{
\href{https://cloud.google.com/translate/docs/}{Google Translation API}
}
\usage{
codebook_trans(cbn)
}
\arguments{
\item{cbn}{File name of the codebook data in the folder}
}
\value{
English and Chinese version of codebook contaning all the question
and answer labels in the survey. Save into data folder.
}
\description{
Translate the codebook data based on \code{\link{trans_data}}
in the data folder, into English, via Googld Translation API.
Save formatted codebook to data folder.
}
