% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_data.R
\name{dta2r}
\alias{dta2r}
\title{Transform General Social Survey Source Data to rda format.}
\source{
\href{http://cnsda.ruc.edu.cn}{Chinese National Survey Data Archive}

\href{http://cgss.ruc.edu.cn}{Chinese General Social Survey Project}
}
\usage{
dta2r(var.name, code.name)
}
\arguments{
\item{var.name}{File name of the source data to be read}

\item{code.name}{File name of the codebook data to be written}
}
\value{
.rda formate of source daata and codebook, save into data folder.
}
\description{
With the source file name of General Social Survey,
it read the stata format source data, save it to .rda format.
And it extract the question and the answer labels, store those
metadata to codebook in a .rda format, for convenient reference
when using the General Social Survey data. \code{\link{trans_codebook}}
would translate the codebook and organize the metadata into both
English and Chinese Version.
}
