% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_variables.R
\name{ques_search_en}
\alias{ques_search_en}
\title{Search variables in the English Codebook
with keywords in the question.}
\usage{
ques_search_en(keyword, codebook)
}
\arguments{
\item{keyword}{Character string of keyword to match the question.}

\item{codebook}{the codebook}
}
\value{
a tibble including\code{index} the index of matched question,
\code{question}, the complete question.
}
\description{
Function to find variables with English keyworks
in question among the Chinese General Social Survey.
}
\examples{
data(codebook2003_en)
ques_search_en("trust", codebook2003_en)

}
