% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_codebook.R
\name{save_trans_code}
\alias{save_trans_code}
\title{Save English and Chinese Codebook}
\usage{
save_trans_code(cbn)
}
\arguments{
\item{cbn}{File name of the codebook data in the folder}
}
\value{
save to data folder
}
\description{
Save cleaned codebook to data folder
}
